let CpgLoader produce a semantic cpg (by already running the enhancements)
