+++ cpgqueryingtests/src/test/scala/io/shiftleft/cpgqueryingtests/codepropertygraph/CpgTestFixture.scala
@@ -1,7 +1,7 @@
 package io.shiftleft.cpgqueryingtests.codepropertygraph
 
 import gremlin.scala._
-import io.shiftleft.queryprimitives.CpgLoader
+import io.shiftleft.CpgLoader
 
 case class CpgTestFixture(projectName: String) {
   lazy val cpg = CpgLoader.loadCodePropertyGraph(s"resources/cpgs/$projectName/cpg.bin.zip")
